fix: import and speedup process with subprocess `communicate()`

+++ voice_chat.py
@@ -10,19 +10,20 @@ from pyrogram.types import Message
 from utils.misc import modules_help, prefix
 from utils.scripts import format_exc, import_library, with_reply, restart
 
-pytgcalls = import_library("pytgcalls","pytgcalls==3.0.0.dev24")
-yt_dlp = import_library("yt_dlp")
+import_library("pytgcalls", "pytgcalls==3.0.0.dev24")
+import_library("yt_dlp")
+
 from pytgcalls import GroupCallFactory
 
-group_call = None
+GROUP_CALL = None
 
 
 def init_client(func):
     async def wrapper(client, message):
-        global group_call
-        if not group_call:
-            group_call = GroupCallFactory(client).get_file_group_call()
-            group_call.enable_logs_to_console = False
+        global GROUP_CALL
+        if not GROUP_CALL:
+            GROUP_CALL = GroupCallFactory(client).get_file_group_call()
+            GROUP_CALL.enable_logs_to_console = False
 
         return await func(client, message)
 
@@ -32,7 +33,7 @@ def init_client(func):
 @Client.on_message(filters.command("play", prefix) & filters.me)
 @with_reply
 async def start_playout(_, message: Message):
-    if not group_call:
+    if not GROUP_CALL:
         await message.reply(
             f"<b>You are not joined [type <code>{prefix}join</code>]</b>"
         )
@@ -49,21 +50,20 @@ async def start_playout(_, message: Message):
     ).overwrite_output().run()
     os.remove(audio_original)
     await message.edit(f"<b>Playing {message.reply_to_message.audio.title}</b>...")
-    group_call.input_filename = input_filename
+    GROUP_CALL.input_filename = input_filename
+
 
 @Client.on_message(filters.command("yplay", prefix) & filters.me)
 async def start_ytplayout(_, message: Message):
-    if not group_call:
+    if not GROUP_CALL:
         await message.reply(
             f"<b>You are not joined [type <code>{prefix}join</code>]</b>"
         )
         return
     if len(message.command) > 1:
-      yt_link = message.text.split(maxsplit=1)[1]
+        yt_link = message.text.split(maxsplit=1)[1]
     input_filename = "input.raw"
     await message.edit_text("<b>Downloading...</b>")
-    # audio_original = await message.reply_to_message.download()
-    # await message.edit("<b>Converting..</b>")
     if yt_link.startswith("https://"):
         try:
             yt = await asyncio.create_subprocess_shell(
@@ -71,12 +71,12 @@ async def start_ytplayout(_, message: Message):
                 stdout=PIPE,
                 stderr=PIPE,
             )
-            if yt.returncode !=0:
-                await asyncio.sleep(30)
-                await message.edit_text(f"<b>Playing</b>...")
-                group_call.input_filename = input_filename
+            await yt.communicate()
+            if yt.returncode != 0:
+                raise Exception("FFmpeg conversion failed.")
         except Exception as e:
             await message.edit_text(format_exc(e))
+            return
     else:
         try:
             yt = await asyncio.create_subprocess_shell(
@@ -84,19 +84,22 @@ async def start_ytplayout(_, message: Message):
                 stdout=PIPE,
                 stderr=PIPE,
             )
-            if yt.returncode !=0:
-                await asyncio.sleep(30)
-                await message.edit_text(f"<b>Playing</b>...")
-                group_call.input_filename = input_filename
+            await yt.communicate()
+            if yt.returncode != 0:
+                raise Exception("FFmpeg conversion failed.")
         except Exception as e:
             await message.edit_text(format_exc(e))
+            return
+    await message.edit_text("<b>Playing</b>...")
+    GROUP_CALL.input_filename = input_filename
+
 
 @Client.on_message(filters.command("volume", prefix) & filters.me)
 @init_client
 async def volume(_, message):
     if len(message.command) < 2:
         await message.edit("<b>You forgot to pass volume [1-200]</b>")
-    await group_call.set_my_volume(message.command[1])
+    await GROUP_CALL.set_my_volume(message.command[1])
     await message.edit(
         f"<b>Your volume is set to</b><code> {message.command[1]}</code>"
     )
@@ -110,7 +113,7 @@ async def start(_, message: Message):
         chat_id = int(chat_id)
 
     try:
-        await group_call.start(chat_id)
+        await GROUP_CALL.start(chat_id)
         await message.edit("<b>Joined VC successfully!</b>")
     except Exception as e:
         await message.edit(f"<b>An unexpected error has occurred: <code>{e}</code></b>")
@@ -120,8 +123,9 @@ async def start(_, message: Message):
 @init_client
 async def stop(_, message: Message):
     try:
-        # os.remove('input.raw')
-        await group_call.stop()
+        if os.path.exists("input.raw"):
+            os.remove("input.raw")
+        await GROUP_CALL.stop()
         await message.edit("<b>Leaving successfully!</b>")
     except Exception as e:
         await message.edit(
@@ -135,36 +139,37 @@ async def stop(_, message: Message):
 @Client.on_message(filters.command("stop", prefix) & filters.me)
 @init_client
 async def stop_playout(_, message: Message):
-    group_call.stop_playout()
-    os.remove('input.raw')
+    GROUP_CALL.stop_playout()
+    if os.path.exists("input.raw"):
+        os.remove("input.raw")
     await message.edit("<b>Stopped successfully!</b>")
 
 
 @Client.on_message(filters.command("vmute", prefix) & filters.me)
 @init_client
 async def mute(_, message: Message):
-    group_call.set_is_mute(True)
+    GROUP_CALL.set_is_mute(True)
     await message.edit("<b>Sound off!</b>")
 
 
 @Client.on_message(filters.command("vunmute", prefix) & filters.me)
 @init_client
 async def unmute(_, message: Message):
-    group_call.set_is_mute(False)
+    GROUP_CALL.set_is_mute(False)
     await message.edit("<b>Sound on!</b>")
 
 
 @Client.on_message(filters.command("pause", prefix) & filters.me)
 @init_client
 async def pause(_, message: Message):
-    group_call.pause_playout()
+    GROUP_CALL.pause_playout()
     await message.edit("<b>Paused!</b>")
 
 
 @Client.on_message(filters.command("resume", prefix) & filters.me)
 @init_client
 async def resume(_, message: Message):
-    group_call.resume_playout()
+    GROUP_CALL.resume_playout()
     await message.edit("<b>Resumed!</b>")
